Configured Pirate Cove Sign Materials
Was originally going to do vertex painting but that found that with the time I have it would be easier to give a material graph set up a try. Doesn't look the best but I'm happy I learned the process at least. Hopefully can add more parameters to control before deadline